Results on test with model: TP : 3151, FP : 50, TN : 50312, FN : 2184, FPR : 0.000993, FNR : 0.409372, PRC : 0.984380, RCL : 0.590628, F1 : 0.738285

+++ experiment/src/data_loader.py
@@ -164,5 +164,5 @@ def eval_with_model(df: pd.DataFrame, df_missing: pd.DataFrame, predictions: np.
 
 
 def get_y_labels(df: pd.DataFrame) -> np.ndarray:
-    true_cases = np.array(df["GroundTruth"], dtype=np.float16)
+    true_cases = np.array(df["GroundTruth"], dtype=np.float32)
     return true_cases
